feat: add crud todo
The CRUD todo feature is the proof of work, which this project is.

The project uses Koa framework, Knex as the query builder, and coded in functional programming paradigm.

This project gives the basic idea of what the code in functional programming looks like, some functional programming design pattern, and serves as the working example of Koa framework implementation.